Improve interop with custom #[test] attributes
As it turns out the attribute, as-is, doesn't play very well with custom
#[test] attributes. The problem is that, because we introduce a module
hierarchy reflecting the tags in use and then use glob imports to pull
in symbols from the "original" scope, Rust fails to disambiguate between
a custom #[test] attribute and the one pull in via the standard prelude.
Admittedly, it's not 100% clear how such a glob import screws us over.

This change works around any such issues, by rewriting #[test]
attributes to #[self::test] and making sure that ::core::prelude::v1::*
is glob imported explicitly as well at the outer-most level, to ensure
that there definitely is a #[test] attribute present, even if the user
didn't include one themselves. Once again, it is not entirely clear why
the standard prelude is not sufficient (it may or may not have to do
with the sparsely documented #[prelude_import]).